move internal.h to expat_internal.h to avoid collision with vmprof file
