+++ pypy/module/pyexpat/src/expat/expat_internal.h
